Moved Gemini code from prompt.go to gemini.go
